Unit 3
This homepage is a work in progress but I enjoyed learning new
concepts. I am still deciding on what color scheme I want to use so for
now my background is white. I do have an idea of what kind of format I
would like for the website and I look forward to the future units!